reorg entire proxy
