+++ magic_bridge/ic/src/dip20_proxy/src/api/init.rs
@@ -1,0 +1,8 @@
+use ic_kit::{ic, macros::*};
+
+use crate::proxy::STATE;
+
+#[init]
+pub fn init() {
+    STATE.with(|s| s.controllers.borrow_mut().push(ic::caller()));
+}
