+++ magic_bridge/ic/src/dip20_proxy/src/api/mod.rs
@@ -1,0 +1,6 @@
+mod burn;
+mod handle_message;
+mod init;
+mod mint;
+mod upgrade;
+mod withdraw;
