+++ magic_bridge/ic/src/dip20_proxy/src/api/withdraw.rs
@@ -1,0 +1,4 @@
+/// Allow user to widthdraw any left over tokens
+pub fn widthdraw() {
+    todo!()
+}
