+++ magic_bridge/ic/src/dip20_proxy/src/common/mod.rs
@@ -1,0 +1,4 @@
+pub mod dip20;
+pub mod tera;
+pub mod types;
+pub mod utils;
