+++ magic_bridge/ic/src/dip20_proxy/src/lib.rs
@@ -1,16 +1,16 @@
-mod dip20;
+mod api;
+mod common;
 mod proxy;
-mod tera;
-mod types;
-mod upgrade;
-mod utils;
 
 #[cfg(any(target_arch = "wasm32", test))]
 fn main() {}
 
 #[cfg(not(any(target_arch = "wasm32", test)))]
 fn main() {
+    use common::types::*;
     use ic_kit::candid;
+    use ic_kit::candid::Nat;
+    use ic_kit::Principal;
 
     candid::export_service!();
     std::print!("{}", __export_service());
